requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
phonenumbers==8.13.33
prompt-toolkit==3.0.36
python-dotenv==1.0.1
questionary==2.0.1
requests==2.31.0
rocketreach==2.1.6
sniffio==1.3.1
truecallerpy==1.0.3
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
